requirements.txt
asyncio==3.4.3
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.9.4
gevent==1.4.0
greenlet==0.4.15
idna==2.8
packaging==19.0
pip-review==1.0
psycopg2==2.8.3
pycparser==2.19
pyparsing==2.4.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
websocket-client==0.56.0
gunicorn==19.7.1

